+++ CHANGELOG.rst
@@ -3,6 +3,30 @@ Change Log
 ###########
 
 
+Unreleased
+**********
+
+New
+---
+* read_nextxy method to read binary nextxy data
+
+Bugfix
+------
+* Support -9 (river outlet at ocean) and -10 (inland river pit) pit values for nextxy data
+
+
+0.5.4  (18-Jan-2022)
+********************
+
+Improved
+---------
+* prioritize non-boundary cells with same elevation over boundary cells in dem.fill_depressions #17
+
+Bugfix
+------
+* fix dem_adjust method #16
+
+
 0.5.3  (18-Nov-2021)
 ********************